hasura_tooling_cli/pyproject.toml
[tool.poetry]
name = "hasura_tooling_cli"
version = "0.1.3"
description = ""

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3"
psycopg2-binary = "2.8.3"
click = "^7.0"
six = "^1.16.0"
hasura-tooling = "^0.10.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.782"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
